chatbot/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
google-generativeai==0.5.4
python-multipart==0.0.9
